Alteração em todas as migrations

+++ database/migrations/2016_08_04_002444_create_estabelecimento_entregas_table.php
@@ -18,7 +18,7 @@ class CreateEstabelecimentoEntregasTable extends Migration
             $table->integer('estabelecimento_id')->unsigned();
             $table->foreign('estabelecimento_id')->references('id')->on('estabelecimentos');
 			$table->integer('tempo_entrega')->default(1);
-			$table->decimal('faixa_preco')->default(1);
+			$table->integer('faixa_preco')->default(1);
 			$table->integer('tipo_pagamento')->default(1);
 			$table->integer('tipo_entrega')->default(1);
